[UPDATE] Mise à jour des functions de la classe Checksum

+++ src/Checksum.php
@@ -2,36 +2,40 @@
 
 namespace SafePHP;
 
+use ErrorException;
 use SafePHP\GLOBALS\Globals;
+
 
 class Checksum {
 
-    public static function createCheckSum($file){
+    public static function createCheckSum($file) : bool | string {
         return hash_file("SHA256", $file, false);
     }
 
-    public static function HasChanged($aFile, $name) : bool { //Supprimer oldchecksum et le chercher
-        if(self::exist($aFile)){
-            $oldChecksum = json_decode($aFile);
+    public static function hasChanged($aFile) : bool {
+        if (self::exist($aFile) === true) {
+            $fileContent = file_get_contents(Globals::$checksumDir . $aFile);
+            $file = json_decode($fileContent, true);
+            $oldChecksum = $file["checksum"];
             $actualChecksum = self::createCheckSum($aFile);
-            if($actualChecksum !== $oldChecksum) {
-                return true;
+            if ($actualChecksum !== $oldChecksum) {
+                throw new ErrorException("Both are not the same ! \n Checksum of the file : " . $oldChecksum . "\nNew checksum :" . $actualChecksum);
             } else {
                 return false;
             }
         } else {
-            return false;
+            throw new ErrorException("This file does not exist in Checksum folder!");
         }
     }
 
-    public static function exist(string $filename){
+    public static function exist(string $filename) : bool{
         $checksumFolder = Globals::$checksumDir;
-        var_dump($checksumFolder);
         $listOfChecksum = scandir($checksumFolder);
-        if($listOfChecksum === false){
+        $fullFileName = $filename . ".json";
+        if ($listOfChecksum === false) {
             return false;
         } else {
-            if(in_array($filename, $listOfChecksum)){
+            if (in_array($fullFileName, $listOfChecksum)) {
                 return true;
             } else {
                 return false;
@@ -39,13 +43,20 @@ class Checksum {
         }
     }
 
-    public static function addToChecksum($file, string $name){
+    public static function addToChecksum($file, string $name): bool | int{
         $checksumExtension = ".json";
         $checkSumDir = Globals::$checksumDir;
         $hashFile = self::createCheckSum($file);
-        
-        $newFile = ["name" => $name, "checksum" => $hashFile];
+        if($hashFile === false){
+            throw new ErrorException("The file specified was not found !");
+        }
+
+        $newFile = ["name" => $name, "Path" => $file, "checksum" => $hashFile];
         $jsonFileContent = json_encode($newFile, JSON_PRETTY_PRINT);
-        return file_put_contents($checkSumDir . $name . $checksumExtension, $jsonFileContent);
+        if(self::exist($name)){
+            throw new ErrorException("This file name is already taken, please use another one !");
+        } else {
+            return file_put_contents($checkSumDir . $name . $checksumExtension, $jsonFileContent);
+        }
     }
 }
